Move bean configuration to corresponding modules
