+++ backend/backend-application/src/main/java/eu/socialedge/hermes/backend/application/config/DomainConfig.java
@@ -15,33 +15,11 @@
 package eu.socialedge.hermes.backend.application.config;
 
 import eu.socialedge.hermes.backend.schedule.infrasturcture.config.ScheduleConfiguration;
-import eu.socialedge.hermes.backend.transit.domain.DistanceAwareSegmentFactory;
-import eu.socialedge.hermes.backend.transit.domain.GMapsTravelDistanceMeter;
-import eu.socialedge.hermes.backend.transit.domain.TravelDistanceMeter;
-import eu.socialedge.hermes.backend.transit.domain.repository.handlers.RouteSegmentLengthHook;
 import eu.socialedge.hermes.backend.transit.infrastructire.config.TransitConfiguration;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 
 @Configuration
 @Import({TransitConfiguration.class, ScheduleConfiguration.class})
 public class DomainConfig {
-
-    @Bean
-    @Value("${ext.google-maps.api}")
-    public TravelDistanceMeter travelDistanceMeter(String apiKey) {
-        return new GMapsTravelDistanceMeter(apiKey);
-    }
-
-    @Bean
-    public DistanceAwareSegmentFactory segmentFactory(TravelDistanceMeter travelDistanceMeter) {
-        return new DistanceAwareSegmentFactory(travelDistanceMeter);
-    }
-
-    @Bean
-    public RouteSegmentLengthHook routeSegmentLengthHook(DistanceAwareSegmentFactory distanceAwareSegmentFactory) {
-        return new RouteSegmentLengthHook(distanceAwareSegmentFactory);
-    }
 }
