+++ backend/backend-transit/src/main/java/eu/socialedge/hermes/backend/transit/infrastructire/config/TransitConfiguration.java
@@ -15,9 +15,14 @@
 
 package eu.socialedge.hermes.backend.transit.infrastructire.config;
 
+import eu.socialedge.hermes.backend.transit.domain.DistanceAwareSegmentFactory;
+import eu.socialedge.hermes.backend.transit.domain.GMapsTravelDistanceMeter;
+import eu.socialedge.hermes.backend.transit.domain.TravelDistanceMeter;
+import eu.socialedge.hermes.backend.transit.domain.repository.handlers.RouteSegmentLengthHook;
 import eu.socialedge.hermes.backend.transit.infrastructure.persistence.QuantityConverters;
 import lombok.val;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
@@ -64,5 +69,21 @@ public class TransitConfiguration {
         return new CustomConversions(asList(
             new QuantityConverters.StringToQuantity(),
             new QuantityConverters.QuantityToStringConverter()));
+    }
+
+    @Bean
+    @Value("${ext.google-maps.api}")
+    public TravelDistanceMeter travelDistanceMeter(String apiKey) {
+        return new GMapsTravelDistanceMeter(apiKey);
+    }
+
+    @Bean
+    public DistanceAwareSegmentFactory segmentFactory(TravelDistanceMeter travelDistanceMeter) {
+        return new DistanceAwareSegmentFactory(travelDistanceMeter);
+    }
+
+    @Bean
+    public RouteSegmentLengthHook routeSegmentLengthHook(DistanceAwareSegmentFactory distanceAwareSegmentFactory) {
+        return new RouteSegmentLengthHook(distanceAwareSegmentFactory);
     }
 }
